changes made

+++ frontend/src/service/roleProvider.tsx
@@ -1,31 +1,48 @@
 import axios from "axios";
 import { RoleDataType } from "../types/roleDataType";
+import * as mockService from "./mockService";
 
 const roleServiceUrl = "http://localhost:5000/api/employee-role";
+const mockEnabled = false;
 
 const getRoles = async () => {
-  const response = await axios.get(`${roleServiceUrl}/`);
-  return response.data;
+  if (mockEnabled) return mockService.getRoles();
+  else {
+    const response = await axios.get(`${roleServiceUrl}/`);
+    return response.data;
+  }
 };
 
 const getRole = async (roleId: any) => {
-  const response = await axios.get(`${roleServiceUrl}/${roleId}`);
-  return response.data;
+  if (mockEnabled) return mockService.getRoles();
+  else {
+    const response = await axios.get(`${roleServiceUrl}/${roleId}`);
+    return response.data;
+  }
 };
 
 const addRoles = async (data: RoleDataType) => {
-  const response = await axios.post(`${roleServiceUrl}/`, data);
-  return response.data;
+  if (mockEnabled) return mockService.getRoles();
+  else {
+    const response = await axios.post(`${roleServiceUrl}/`, data);
+    return response.data;
+  }
 };
 
 const editRole = async (roleId: any, data: RoleDataType) => {
-  const response = await axios.patch(`${roleServiceUrl}/${roleId}`, data);
-  return response.data;
+  if (mockEnabled) return mockService.getRoles();
+  else {
+    const response = await axios.patch(`${roleServiceUrl}/${roleId}`, data);
+    return response.data;
+  }
 };
 
 const deleteRole = async (roleId: any) => {
-  const response = await axios.delete(`${roleServiceUrl}/${roleId}`);
-  return response.data;
+  if (mockEnabled) return mockService.getRoles();
+  else {
+    const response = await axios.delete(`${roleServiceUrl}/${roleId}`);
+    return response.data;
+  }
 };
 
 export default {
